+++ ANPAHP/rag/logic/__init__.py
@@ -14,11 +14,11 @@ For example, you can create a new file under `ANPAHP.rag.logic` called `my_logic
 Here is the minimal content of the file:
 
 ```
+from chromadb import Collection
 from langchain.llms.base import LLM
 from langchain_core.prompts import ChatPromptTemplate
-from langchain_chroma import Chroma
 
-def rag_logic(llm: LLM, vector_store: Chroma, user_query: str) -> tuple[str, list[str]]:
+def rag_logic(llm: LLM, vector_store: Collection, user_query: str) -> tuple[str, list[str]]:
     return "test", ["This is test logic A", "This is test logic B"]
 ```